Latest

+++ actors/git.php
@@ -66,11 +66,15 @@ class git extends \update\updater implements actorsInterface {
 
 		$this->_output(parent::H3,"Self Update");
 
-		$this->_output(parent::EXEC,"INIT",$this->_exec('git init',$dir));
-		$this->_output(parent::EXEC,"STASH",$this->_exec('git reset --hard HEAD',$dir));
+		if (is_file($dir."Updater".DIRECTORY_SEPARATOR.".git")){
+			$this->_output(parent::EXEC,"SELF UPDATING",$this->_exec('git submodule update --recursive --remote',$dir));
+		} else {
+			$this->_output(parent::EXEC,"INIT",$this->_exec('git init',$dir));
+			$this->_output(parent::EXEC,"STASH",$this->_exec('git reset --hard HEAD',$dir));
+			$this->_output(parent::EXEC,"SELF UPDATING",$this->_exec('git pull https://'.$self['git']['path'] .' ' . $self['git']['branch']."",$dir));
+		}
 
 
-		$this->_output(parent::EXEC,"SELF UPDATING",$this->_exec('git pull https://'.$self['git']['path'] .' ' . $self['git']['branch']."",$dir));
 
 
 	}
